990825-20:25

+++ HandCarftBaseServer/ServiceProvider/Zarinpal/ZarinPalVerifyResponse.cs
@@ -9,7 +9,7 @@ namespace HandCarftBaseServer.ServiceProvider.ZarinPal
     {
         public int code { set; get; }
         public int merchacodent_id { set; get; }
-        public int ref_id { set; get; }
+        public long ref_id { set; get; }
         public string card_pan { set; get; }
         public string card_hash { set; get; }
         public string fee_type { set; get; }
